영화 숫자가 유효하는지 확인하고 오류 메세지 출력과 재입력을 요구하는 기능 추가
